creating separate EngineSchematic class
