+++ src/utils/stringCheckers.ts
@@ -1,0 +1,11 @@
+export const isNumber = (value: string): boolean => !isNaN(Number(value));
+// TODO: add more characters
+export const isSymbol = (value: string): boolean =>
+  value.match(/^[@#$%&*-+=//]*$/) ? true : false;
+/**
+ * Day Three specific checkers
+ */
+export const dayThree = {
+  isEngineSymbol: (value: string): boolean => !isNumber(value) && value !== ".",
+  isGear: (value: string): boolean => value === "*",
+};
